change the return type of getAccountID() from HashMap<String,Double> to HashMap<String,Long>. We previously changed the HashMap value to Double because of a json problem, which we've already fixed. But now since msn is not dealing with any json, we'll change it back, since Long makes more sense for AccountID.
git-svn-id: file:///semplestb/final_repo@3197 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SEMplestNIOProtocol/src/semplest/services/client/interfaces/MsnAdcenterServiceInterface.java
@@ -52,7 +52,7 @@ public interface MsnAdcenterServiceInterface extends ServiceInitialize {
 	 */
 	MsnManagementIds createAccount(SemplestString name) throws Exception;  	
 	Account getAccountById(Long accountId) throws Exception;	
-	HashMap<String,Double> getAccountIDs() throws Exception;
+	HashMap<String,Long> getAccountIDs() throws Exception;
 	
 	// ==================================
 	// GeoTargeting Methods
